* StateOfCharge on PlotLineItem * SecondaryDimension for PlotView rendering * Speed and StateOfCharge PlotLines not needed anymore as implicit from stored data * Restore old Data to new Format

+++ automotive/src/main/java/com/ixam97/carStatsViewer/plot/PlotPoint.kt
@@ -7,4 +7,4 @@ class PlotPoint(
 class PlotLineItemPoint(
     val x: Float,
     val y: PlotLineItem,
-    val group: Long)
+    val group: Float)
